+++ CHANGES.md
@@ -2,6 +2,9 @@
 
 ## Next
 
+- Feature: Added command tests
+- Feature: Created show command
+- Feature: Created fixtures
 - Minor: Added GitHub Actions Badge
 - Feature: Added GitHub Actions directory
 - Cleanup: Removed examples directory